Halve the panel width so the eleven-year study fits in memory
The first attempt at the pre-registered run was OOM-killed during feature
assembly. At 8 million rows and ~150 float64 columns the merges inside
assemble() materialise several tens of gigabytes, and the machine has ten.

Float columns are now downcast to float32 in the assembler and in the
training matrix. This is a memory change, not a modelling one, and the
distinction is worth stating precisely because the run it enables is
pre-registered: LightGBM does not split on raw floats. It bins each feature
into at most max_bin buckets -- 255 by default -- and splits on bin edges.
Float32 carries about seven significant decimal digits, finer than any bin
boundary by orders of magnitude, so binning, splits and predictions are
unchanged. The only thing that would threaten is a feature carrying
information beyond the seventh significant digit, and there is none: these
are returns, volatilities, decayed event counts and dollar volumes. Integer
and boolean columns are left alone, since downcasting a count changes what
it can represent.

Intermediates are also freed as soon as they are consumed -- the market and
event frames inside assemble, and the 2.5-million-row event frame in the
script once features exist, which nothing downstream reads.

+++ src/iai/moonshot.py
@@ -94,7 +94,11 @@ def train_moonshot(
         raise ValueError("no rows after joining panel to spike labels")
 
     feats = [c for c in panel.columns if c not in ("date", "ticker")]
-    X = df[feats].astype("float64")
+    # float32, for the same reason the assembler downcasts: LightGBM bins to at
+    # most 255 buckets, so the extra mantissa is discarded before a single split
+    # is chosen, and at eight million rows the wider copy is the difference
+    # between running and being OOM-killed.
+    X = df[feats].astype("float32")
     y_up = df["spike"].astype(int)
     y_dn = df["exit_reason"].isin(STOP_REASONS).astype(int)
